GeoJSONDataProvider is now a factory that knows how to generate dataProviders for each type of dataview. Dataviews use a specific type of dataProvider.
