+++ src/dataviews/dataview-model-base.js
@@ -48,6 +48,10 @@ module.exports = Model.extend({
       this.filter.set('dataviewId', this.id);
     }
 
+    var dataProvider = this.layer.getDataProvider();
+    if (dataProvider) {
+      this._dataProvider = dataProvider.createDataProviderForDataview(this);
+    }
     this._initBinds();
     this._updateBoundingBox();
   },
@@ -56,17 +60,9 @@ module.exports = Model.extend({
     this.listenTo(this._windshaftMap, 'instanceCreated', this._onNewWindshaftMapInstance);
     this.listenTo(this.layer, 'change:visible', this._onLayerVisibilityChanged);
 
-    var dataProvider = this.layer.getDataProvider();
-    if (dataProvider) {
-      this.listenToOnce(dataProvider, 'featuresChanged', function () {
-        this.listenTo(this._map, 'change:center change:zoom', _.debounce(this._onMapBoundsChanged.bind(this), BOUNDING_BOX_FILTER_WAIT));
-        this.on('change:boundingBox', function () {
-          if (this._shouldFetchOnBoundingBoxChange()) {
-            this.fetch();
-          }
-        }, this);
-      });
-      this.listenTo(dataProvider, 'featuresChanged', this._onDataProviderChanged);
+    if (this._dataProvider) {
+      this.listenToOnce(this._dataProvider, 'dataChanged', this._onChangeBinds, this);
+      this.listenTo(this._dataProvider, 'dataChanged', this._onDataProviderChanged);
     } else {
       this.listenToOnce(this, 'change:url', function () {
         this.fetch({
@@ -79,24 +75,16 @@ module.exports = Model.extend({
     }
   },
 
-  _onDataProviderChanged: function (features) {
-    var dataProvider = this.layer.getDataProvider();
-    var data = dataProvider.generateDataForDataview(this, features);
-
-    // TODO: this.parse() is doing some magic (like resetting this._data)
-    // in some cases (category and histogram) and that's why we use it here.
-    // If that wasn't the case, data could be generated by the dataProvider in
-    // the format that the dataview expects.
-    this.set(this.parse(data));
+  _onDataProviderChanged: function () {
+    this.set(this.parse(this._dataProvider.getData()));
   },
 
   /**
    * @private
    */
   _onFilterChanged: function (filter) {
-    var dataProvider = this.layer.getDataProvider();
-    if (dataProvider) {
-      dataProvider.applyFilter(this.get('column'), filter);
+    if (this._dataProvider) {
+      this._dataProvider.applyFilter(filter);
     } else {
       this._reloadMap();
     }
@@ -200,10 +188,8 @@ module.exports = Model.extend({
 
   fetch: function (opts) {
     opts = opts || {};
-    var dataProvider = this.layer.getDataProvider();
-    if (dataProvider) {
-      var data = dataProvider.generateDataForDataview(this, dataProvider.getData());
-      this.set(this.parse(data));
+    if (this._dataProvider) {
+      this.set(this.parse(this._dataProvider.getData()));
       this.trigger('sync');
     } else {
       this.trigger('loading', this);
